Back-End/src/main/java/com/example/BackEnd/Controller/OrderController.java
package com.example.BackEnd.Controller;

import com.example.BackEnd.DTO.OrderDTO;
import com.example.BackEnd.DTO.ResponseDTO;
import com.example.BackEnd.Service.OrderService;
import com.example.BackEnd.util.VarList;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

@RestController
@RequestMapping(value = "/api/v1/order")
public class OrderController {
    @Autowired
    private OrderService orderService;
    @Autowired
    private ResponseDTO responseDTO;


    @PostMapping(value = "/placeOrder")
    public ResponseEntity place_order(@RequestBody OrderDTO orderDTO){

        try {
            String resp = orderService.place_order(orderDTO);
            ResponseDTO responseDTO = new ResponseDTO();

            if (resp.equals("00")) {
                responseDTO.setCode(VarList.RSP_SUCCESS);
                responseDTO.setMessage("Success");
                responseDTO.setContent(orderDTO);
                return new ResponseEntity(responseDTO, HttpStatus.ACCEPTED);
            } else if (resp.equals("01")) {
                responseDTO.setCode(VarList.RSP_NO_DATA_FOUND);
                responseDTO.setMessage(null);
                responseDTO.setContent(orderDTO);
                return new ResponseEntity(responseDTO, HttpStatus.BAD_REQUEST);
            } else {
                responseDTO.setCode(VarList.RSP_FAIL);
                responseDTO.setMessage("Error");
                responseDTO.setContent(null);
                return new ResponseEntity(responseDTO, HttpStatus.BAD_REQUEST);
            }

        } catch (Exception x) {
            responseDTO.setCode(null);
            responseDTO.setMessage(x.getMessage());
            responseDTO.setContent(null);
            return new ResponseEntity(responseDTO, HttpStatus.INTERNAL_SERVER_ERROR);
        }

    }
}
